+++ playground/pages/index.vue
@@ -121,9 +121,21 @@ const onInitialized = (NuxtEditorJs) => {
   console.log(NuxtEditorJs)
 }
 
+// https://github.com/pavittarx/editorjs-html/tree/master#extend-for-custom-blocks
+function customParser (block) {
+  return `<custom-tag> ${block.data.text} </custom-tag>`
+}
+const configParseHtml = {
+  custom: customParser
+}
+const { parse, parseBlock, parseStrict, validate, resultString } = useParseHtml(configParseHtml)
+
 watch(dat, (value) => {
   // eslint-disable-next-line no-console
   console.log(value)
+  // using custom parse
+  parse(value)
+  console.log(resultString())
 }, { deep: true, immediate: true })
 </script>
