+++ src/runtime/plugin.client.ts
@@ -1,6 +1,6 @@
 // @ts-ignore
 import { defineNuxtPlugin } from '#app'
-import { NuxtEditorJs } from './composables/Editorjs.client'
+import { NuxtEditorJs } from './components/Editorjs.client'
 
 export default defineNuxtPlugin((nuxtApp) => {
   nuxtApp.vueApp.component('NuxtEditorJs', NuxtEditorJs)
